Update to usize, remove wasm bindgens (use a wrapper lol)

+++ src/datastruct.rs
@@ -1,17 +1,15 @@
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 #[derive(Serialize, Deserialize)]
 pub struct PathOpt {
-    jump: u16,
+    jump: usize,
     text: String,
 }
 
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 impl PathOpt {
     /// Gets the place where the path is meant to jump
-    pub fn get_jump(&self) -> u16 {
+    pub fn get_jump(&self) -> usize {
         self.jump
     }
     /// Gets the text for the option
@@ -20,14 +18,12 @@ impl PathOpt {
     }
 }
 
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 #[derive(Serialize, Deserialize)]
 pub struct Path {
     text: String,
     options: Vec<PathOpt>,
 }
 
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 impl Path {
     /// Gets the text for the path
     pub fn get_text(&self) -> &String {
@@ -39,16 +35,14 @@ impl Path {
     }
 }
 
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 #[derive(Serialize, Deserialize)]
 pub struct Game {
     name: String,
     author: String,
     slug: String,
-    paths: HashMap<u16, Path>,
+    paths: HashMap<usize, Path>,
 }
 
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 impl Game {
     /// Getter for the name of the game
     pub fn get_name(&self) -> &String {
@@ -63,11 +57,11 @@ impl Game {
         &self.slug
     }
     /// Getter for an arbitrary path
-    pub fn get_path(&self, id: u16) -> &Path {
+    pub fn get_path(&self, id: usize) -> &Path {
         &self.paths.get(&id).unwrap()
     }
     /// Checks to make sure that a path exists
-    pub fn check_path(&self, path: &u16) -> bool {
+    pub fn check_path(&self, path: &usize) -> bool {
         self.paths.contains_key(path)
     }
 }
